+++ src/components/button-bar/button-bar.test.js
@@ -1,6 +1,5 @@
 import React from "react";
-import ButtonBar from ".";
-import Button from "../button";
+import { Default as ButtonBarCustom } from "./button-bar-test.stories";
 import {
   BUTTON_BAR_SIZES,
   BUTTON_BAR_ICON_POSITIONS,
@@ -9,29 +8,8 @@ import {
 import { buttonDataComponent } from "../../../cypress/locators/button";
 
 import { icon } from "../../../cypress/locators";
-import { positionOfElement, keyCode } from "../../../cypress/support/helper";
 import { useJQueryCssValueAndAssert } from "../../../cypress/support/component-helper/common-steps";
 import CypressMountWithProviders from "../../../cypress/support/component-helper/cypress-mount";
-
-const ButtonBarCustom = ({ onClick, ...props }) => {
-  return (
-    <ButtonBar onClick={onClick} {...props} ml={2} mt={2}>
-      <Button>Button 1</Button>
-      <Button>Button 2</Button>
-      <Button>Button 3</Button>
-    </ButtonBar>
-  );
-};
-
-const ButtonBarIconPosition = ({ iconPositionProp, ...props }) => {
-  return (
-    <ButtonBar {...props} size={BUTTON_BAR_SIZES[0]} ml={2} mt={2}>
-      <Button iconType="csv">{iconPositionProp} Button</Button>
-      <Button iconType="pdf">{iconPositionProp} Button</Button>
-      <Button iconType="delete">{iconPositionProp} Button</Button>
-    </ButtonBar>
-  );
-};
 
 context("Test for Button-Bar component", () => {
   describe("check props for Button-Bar component", () => {
@@ -54,7 +32,7 @@ context("Test for Button-Bar component", () => {
       "should set position to %s for icon in a Button-Bar",
       (iconPosition, margin) => {
         CypressMountWithProviders(
-          <ButtonBarIconPosition iconPosition={iconPosition} />
+          <ButtonBarCustom iconPosition={iconPosition} />
         );
 
         icon().should("have.css", `margin-${margin}`, "8px");
@@ -70,63 +48,33 @@ context("Test for Button-Bar component", () => {
           useJQueryCssValueAndAssert($el, "width", 1366);
         });
     });
+  });
 
-    describe("check events for Button-Bar component", () => {
-      let callback;
+  describe("accessibility tests", () => {
+    it.each([BUTTON_BAR_SIZES[0], BUTTON_BAR_SIZES[1], BUTTON_BAR_SIZES[2]])(
+      "should check accessibility for %s size for a Button-Bar",
+      (size) => {
+        CypressMountWithProviders(<ButtonBarCustom size={size} />);
 
-      beforeEach(() => {
-        callback = cy.stub();
-      });
+        cy.checkAccessibility();
+      }
+    );
 
-      it("should call onClick callback when a click event is triggered", () => {
-        CypressMountWithProviders(<ButtonBarCustom onClick={callback} />);
+    it.each([BUTTON_BAR_ICON_POSITIONS[0], BUTTON_BAR_ICON_POSITIONS[1]])(
+      "should check accessibility for %s icon position in a Button-Bar",
+      (iconPosition) => {
+        CypressMountWithProviders(
+          <ButtonBarCustom iconPosition={iconPosition} />
+        );
 
-        buttonDataComponent()
-          .eq(positionOfElement("first"))
-          .click({ force: true })
-          .then(() => {
-            // eslint-disable-next-line no-unused-expressions
-            expect(callback).to.have.been.calledOnce;
-          });
-      });
+        cy.checkAccessibility();
+      }
+    );
 
-      it("should call onBlur callback when a blur event is triggered", () => {
-        CypressMountWithProviders(<ButtonBarCustom onBlur={callback} />);
+    it("should check the accessibility of Button-Bar with full width", () => {
+      CypressMountWithProviders(<ButtonBarCustom fullWidth />);
 
-        buttonDataComponent()
-          .eq(positionOfElement("second"))
-          .focus()
-          .blur({ force: true })
-          .then(() => {
-            // eslint-disable-next-line no-unused-expressions
-            expect(callback).to.have.been.calledOnce;
-          });
-      });
-
-      it("should call onKeyDown callback when a keydown event is triggered", () => {
-        CypressMountWithProviders(<ButtonBarCustom onKeyDown={callback} />);
-
-        buttonDataComponent()
-          .eq(positionOfElement("first"))
-          .trigger("keydown", { force: true }, keyCode("rightarrow"))
-          .then(() => {
-            // eslint-disable-next-line no-unused-expressions
-            expect(callback).to.have.been.calledOnce;
-          });
-      });
-
-      it("should call onFocus callback when a focus event is triggered", () => {
-        CypressMountWithProviders(<ButtonBarCustom onFocus={callback} />);
-
-        buttonDataComponent()
-          .eq(positionOfElement("third"))
-          .focus()
-          .blur({ force: true })
-          .then(() => {
-            // eslint-disable-next-line no-unused-expressions
-            expect(callback).to.have.been.calledOnce;
-          });
-      });
+      cy.checkAccessibility();
     });
   });
 });